chore: Move details to config file
