+++ src/views/Ifos/components/IfoQuestions/config.ts
@@ -1,0 +1,18 @@
+const config = [
+  {
+    title: { id: 999, fallback: "What's the difference between a Basic Sale and Unlimited Sale" },
+    description: {
+      id: 999,
+      fallback: 'TBD',
+    },
+  },
+  {
+    title: { id: 999, fallback: 'TBD' },
+    description: {
+      id: 999,
+      fallback: 'TBD',
+    },
+  },
+]
+
+export default config
